Users Database Connectivity
Database name : optiverse
table name : users

when a user Enters its username and password the website takes him to the home page Otherwise if the username or password doesn't match then a message shows up user is not registered. But there's still few problems with this code that is when you hit Forgot password or terms and conditions buttons it throws an error saying it cannot get those files. So I need to connect these two files just like before I connected the sign in button to another page.